+++ test/Streamly/Test/Serialize/Serializable.hs
@@ -10,8 +10,8 @@ import Test.QuickCheck
     )
 import Test.QuickCheck.Monadic (monadicIO)
 
-import qualified Streamly.Internal.Serialize.FromBytes as Decoder
-import qualified Streamly.Internal.Serialize.ToBytes as Encoder
+import qualified Streamly.Internal.Data.Serialize.Parser as Decoder
+import qualified Streamly.Internal.Data.Serialize.Stream as Encoder
 import qualified Streamly.Data.Stream as Stream
 import Streamly.Internal.Data.Parser (Parser)